Address comments from Jun
